modified:   Makefile modified:   Makefile.body modified:   Makefile.in Added conc_to_pow function and its dependencies. Added compartment_lookup to translate_regulation_metabolites dependencies
new file: conc_to_pow.c
new file: conc_to_pow.h
  Routine and its header to raise a concentration/count to a non-integer power
  also allowing a telescoping (odds) product to be computed.

modified:   README
First line changed to ../README.md

modified:   alloc0.c
version_no => 5098

Most of the changes below have to do with changing stoichiometric coefficients
from being type int64_t or int to type double in the following structs,
abbreviated as coefficients change.

  coeff_sum and rxn_mat_row fields of state struct
  coefficients, and solvent_coefficients fields of reactions_matrix_struct
  coefficients field of molecules_matrix_struct

modified:   alloc2.c
  coefficients change.

modified:   alloc4.c
  coefficients change.

modified:   alloc9.c
  coefficients change.

modified:   bndry_flux_update.c
  coefficients change.

modified:   boltzmann_compress_reactions.c
  coefficients change.

modified:   boltzmann_set_state_ptrs.c
  coefficients change.

modified:   compartment_lookup.c
  documentation change to Called by: line,
  added translate_regulation_betabolites caller.

modified:   compute_dfdke_dfdmu0.c
  coefficients change and adding conc_to_pow calls.

modified:   compute_kss.c
  coefficients change and adding conc_to_pow calls.

modified:   compute_net_lklhd_bndry_flux.c
  coefficients change.

modified:   compute_reaction_dg0.c
  coefficients change.

modified:   count_molecules.c
  coefficients change and added third argument &coeff to is_a_coef call.

modified:   count_molecules_and_cmpts.c
  coefficients change and added third argument &coeff to is_a_coef call.

modified:   cvodes_interface/lr8_approximate_ys0.c
  coefficients change and adding conc_to_pow calls.

modified:   echo_reactions_file.c
  coefficients change.

modified:   form_molecules_matrix.c
  coefficients change.

modified:   is_a_coef.c
modified:   is_a_coef.h
  added third argument to return the value of the coefficient if found.
  Changed to allow floating point coefficients (%le => double).
  This uses sccanf instead of checking for a string of decimal digits.
  Still returns a 1 if the token was a coefficient and 0 if not.

  Does require the coefficient to take up the whole token to avoid
  returning a coefficient from a molcule name that starts with a number ie.
  the token "6-hydroxy-quinone" should return a 0 instead of 1 with coefficient
  value 6.

modified:   molecules_matrix_struct.h
  type of coefficients field changde from int64_t* to double*

modified:   ode23tb/init_base_reactants.c
  coefficients change.

modified:   ode23tb/init_relative_rates.c
  coefficients change.

modified:   ode23tb/lr10_approximate_delta_concs.c
  coefficients change and adding conc_to_pow calls.

modified:   ode23tb/lr11_approximate_delta_concs.c
  coefficients change and adding conc_to_pow calls.

modified:   ode23tb/lr12_approximate_delta_concs.c
  coefficients change and adding conc_to_pow calls.

modified:   ode23tb/lr13_approximate_delta_concs.c
  coefficients change and adding conc_to_pow calls.

modified:   ode23tb/lr14_approximate_delta_concs.c
  coefficients change and adding conc_to_pow calls.

modified:   ode23tb/lr1_approximate_delta_concs.c
  coefficients change.

modified:   ode23tb/lr2_approximate_delta_concs.c
  coefficients change.

modified:   ode23tb/lr3_approximate_delta_concs.c
  coefficients change.

modified:   ode23tb/lr4_approximate_delta_concs.c
  coefficients change and adding conc_to_pow calls.

modified:   ode23tb/lr5_approximate_delta_concs.c
  coefficients change.

modified:   ode23tb/lr6_approximate_delta_concs.c
  coefficients change.

modified:   ode23tb/lr7_approximate_delta_concs.c
  coefficients change and adding conc_to_pow calls.

modified:   ode23tb/lr8_approximate_delta_concs.c
  coefficients change and adding conc_to_pow calls.

modified:   ode23tb/lr8_approximate_jacobian.c
  coefficients change and adding conc_to_pow calls and (coeff => coef).

modified:   ode23tb/lr9_approximate_delta_concs.c
  coefficients change and adding conc_to_pow calls.

modified:   ode23tb/lr_approximate_delta_concs.c
  coefficients change.

modified:   parse_reactions_file.c
  coefficients change, and checking rcompartment for not a single 'V' or 'C'
  name.

modified:   parse_side_line.c
  coefficients change and added lfp initialization.

modified:   print_active_reactions_matrix.c
  coefficients change.

modified:   print_dg0_ke.c
  coefficients change.

modified:   print_reactions_matrix.c
  coefficients change.

modified:   print_reactions_view.c
  coefficients change.

modified:   reaction_struct.h
  coefficient_sum field changed from type int to double.
  padi field for padding removed.

modified:   reactions_matrix_struct.h
  coefficients and solvent_coefficients fields changed from
  int64_t* type to double* type.

modified:   recover_solvent_coefficients.c
  coefficients change and added recovering the reciprocals too.

modified:   rxn_conc_update.c
  coefficients change.

modified:   rxn_count_update.c
  coefficients change.

modified:   rxn_likelihood.c
  coefficients change and adding conc_to_pow calls.

modified:   rxn_likelihood_postselection.c
  coefficients change and adding conc_to_pow calls.

modified:   sbml_interface/sbml_process_list_of_reactions.c
  coefficients change and initialized coefficient to 1.0
  before calling sbml_proces_species_reference_tag.

modified:   sbml_interface/sbml_process_species_reference_tag.c
modified:   sbml_interface/sbml_process_species_reference_tag.h
  coefficients change and that also cause a change in argument list type.

modified:   state_struct.h
  filed coeff_sum and rxn_mat_row changed from int64_t* to double *.

modified:   translate_regulation_metabolites.c
  added include for compartment_lookup header and added state to its
  argument list.

modified:   zero_solvent_coefficients.c
  coefficients change and set recip_coeff of solvent to be 0 as well.

+++ src/alloc4.c
@@ -95,9 +95,9 @@ int alloc4(struct state_struct *state, int setup) {
     } 
   }
   if (success) {
-    ask_for = ((int64_t)(nzr)) * ((int64_t)sizeof(int64_t));
+    ask_for = ((int64_t)(nzr)) * ((int64_t)sizeof(double));
     usage += ask_for;
-    molecules_matrix->coefficients = (int64_t*)calloc(one_l,ask_for);
+    molecules_matrix->coefficients = (double*)calloc(one_l,ask_for);
     if (molecules_matrix->coefficients == NULL) {
       fprintf(stderr,"alloc4: Error unable to allocate %ld bytes for "
 	      "coefficients field in molecules_matrix\n",ask_for);
